requirements.txt
pyaudio
websockets
streamlit
python-dotenv